Rename package from config to clusterdesc
