+++ clusterdesc/linux_distro.go
@@ -1,4 +1,4 @@
-package config
+package clusterdesc
 
 import (
 	"log"
