+++ clusterdesc/master.go
@@ -1,4 +1,4 @@
-package config
+package clusterdesc
 
 import (
 	"fmt"
